Second step workflow modification
Now, the input text for the second step is not the text generated
in the previous step, but a post processed version that only contains
a list of labels to be classified.
The second_iteration function has been modified to implement this
new logic.

Morover, the output of the model is no longer a json, but a csv.
The function output_to_csv has been added to manage this last change.